Decode base64 data in MatchData, disable use_threads in HTML5. (#20)
* Disable HTTPRequest.use_threads in HTML5 exports.

Threads are not (yet) supported in HTML5 exports, and are in any case
not needed since HTTPRequests are always async in browsers.

* Decode base64 data in MatchData.

It's a bit hacky compared to the rest of the code. A cleaner solution
would probably be to add a condition for that in the serializer, but
for just one case it's not worth it. Should reconsider if more base64
messages needs to be handled.

* Update changelog