Wrongly accessing sub-class in a script.
The sub-class names of a script aren't listed as properties of the script rather as constants. Therefore, we can get the class-names defined in the script from the constants list (which is defined in a map/dictionary form).

This was a bug that didn't deserialize the dictionary properly and pack it into the object and so the object returned was empty.